requirements.txt
disnake
flask
pre-commit
python-decouple
cogs
